Fix critical bugs in sleep stage calculator
Fix 4 critical bugs and improve precision:

1. Fix negative sleep latency in all-wake sessions
   - Add guard in _calculate_latencies() for first_sleep_idx == -1
   - Set sleep_latency to entire session duration when no sleep
   - Set wakeup_latency to 0 when no sleep occurred

2. Fix IndexError in cluster calculations
   - Add early return guards in _calculate_wake_cluster_indices()
   - Add early return guards in _calculate_rem_cluster_indices()
   - Return empty results when no sleep occurred (first_sleep_idx == -1)

3. Fix REM double-counting bug
   - Restructure REM cluster logic to count each epoch only once
   - Remove duplicate increment that counted first REM of new clusters twice

4. Fix documentation parameter order
   - Update README.md, calculator/README.md, and calculator.py docstring
   - Correct API call examples to show: calculate(sleep_stages, start_time, end_time)

Improvements:
- Add comprehensive test assertions for latencies and time invariants
- Add new test_time_invariants() test for multiple edge cases
- Remove int() cast in sleep cycle calculation for better precision
- Update sleep_cycle type from int to float in SleepStageClusterMetric

All tests pass (10 passed, 3 skipped).
Both example scripts verified working.

+++ src/pysleep/calculator/types.py
@@ -148,7 +148,7 @@ class SleepStageClusterMetric:
 
     waso_count: int
     longest_waso: int
-    sleep_cycle: Optional[int]
+    sleep_cycle: Optional[float]
     sleep_cycle_count: int
     sleep_cycle_time: List[datetime]
